refactor : 공통 메소드 및 설정 추상메소드로 분리 - 빈 체스판 초기화 메소드 및 초기 설정 부분 추상 메소드로 분리하여 중복 감소
