+++ src/test/java/chess/pieces/KingTest.java
@@ -7,6 +7,7 @@ import chess.Board;
 import chess.ChessGame;
 import chess.ChessView;
 import chess.Rank;
+import chess.TestUtil;
 import chess.pieces.Piece.Type;
 import java.util.ArrayList;
 import java.util.List;
@@ -14,18 +15,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class KingTest {
-    private Board board;
-    private ChessGame chessGame;
-    private ChessView chessView;
+public class KingTest extends TestUtil {
 
-    @BeforeEach
-    void init() {
-        List<Rank> chessBoard = new ArrayList<>();
-        board = new Board(chessBoard);
-        chessGame = new ChessGame(chessBoard);
-        chessView = new ChessView(chessBoard);
-    }
     @Test
     @DisplayName("킹은 8방향으로 한칸씩 이동할 수 있다.")
     void moveKing() {
@@ -64,22 +55,5 @@ public class KingTest {
 
         assertThrows(IllegalArgumentException.class, () ->chessGame.move(sourcePosition, targetPosition));
         assertThat(chessGame.findPiece(sourcePosition)).isEqualTo(Piece.createBlack(Type.KING));
-    }
-
-    private void initEmptyBoardTest() {
-        board.initializeEmpty();
-
-        addPiece("b6", Piece.createBlack(Type.PAWN));
-        addPiece("e6", Piece.createBlack(Type.QUEEN));
-        addPiece("b8", Piece.createBlack(Type.KING));
-        addPiece("c8", Piece.createBlack(Type.ROOK));
-
-        addPiece("g2", Piece.createWhite(Type.PAWN));
-        addPiece("g3", Piece.createWhite(Type.PAWN));
-        addPiece("e1", Piece.createWhite(Type.ROOK));
-        addPiece("f1", Piece.createWhite(Type.KING));
-    }
-    private void addPiece(String position, Piece piece) {
-        chessGame.addPiece(position, piece);
     }
 }
