+++ src/test/java/chess/pieces/QueenTest.java
@@ -3,29 +3,13 @@ package chess.pieces;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import chess.Board;
-import chess.ChessGame;
-import chess.ChessView;
-import chess.Rank;
+import chess.TestUtil;
 import chess.pieces.Piece.Type;
-import java.util.ArrayList;
-import java.util.List;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class QueenTest {
-    private Board board;
-    private ChessGame chessGame;
-    private ChessView chessView;
+public class QueenTest extends TestUtil {
 
-    @BeforeEach
-    void init() {
-        List<Rank> chessBoard = new ArrayList<>();
-        board = new Board(chessBoard);
-        chessGame = new ChessGame(chessBoard);
-        chessView = new ChessView(chessBoard);
-    }
     @Test
     @DisplayName("퀸은 직선과 대각선 방향으로 움직인다.")
     void moveQueen() {
@@ -65,20 +49,5 @@ public class QueenTest {
 
         System.out.println(chessView.showBoard());
     }
-    private void initEmptyBoardTest() {
-        board.initializeEmpty();
 
-        addPiece("b6", Piece.createBlack(Type.PAWN));
-        addPiece("e6", Piece.createBlack(Type.QUEEN));
-        addPiece("b8", Piece.createBlack(Type.KING));
-        addPiece("c8", Piece.createBlack(Type.ROOK));
-
-        addPiece("g2", Piece.createWhite(Type.PAWN));
-        addPiece("g3", Piece.createWhite(Type.PAWN));
-        addPiece("e1", Piece.createWhite(Type.ROOK));
-        addPiece("f1", Piece.createWhite(Type.KING));
-    }
-    private void addPiece(String position, Piece piece) {
-        chessGame.addPiece(position, piece);
-    }
 }
